PRACTICA5/1/web/requirements.txt
Flask==2.0.1
Flask-Login==0.5.0
requests==2.26.0
python-dotenv==0.19.0
Werkzeug==2.0.1 